requirements.txt
PyQt5>=5.15
Pillow>=9.0
pyinstaller>=6.0
